+++ js/keyboard-keys.js
@@ -1,0 +1,4 @@
+const isEscapeKey = (evt) => evt.key === 'Escape';
+const isEnterKey = (evt) => evt.key === 'Enter';
+
+export {isEscapeKey, isEnterKey};
