+++ js/main.js
@@ -1,4 +1,5 @@
-import { renderGallery } from './big-photo.js';
+import { renderGallery } from './photo-handler.js';
 import { photos } from './data.js';
+
 
 renderGallery(photos);
